refactor: change event model
